Refactor to flatten folder structure and make pythonic
