+++ src/electionguard_cli/cli_models/__init__.py
@@ -1,13 +1,10 @@
-from electionguard_cli.cli_models import e2e
+from electionguard_cli.cli_models import cli_election_inputs_base
 from electionguard_cli.cli_models import e2e_build_election_results
 from electionguard_cli.cli_models import e2e_decrypt_results
 from electionguard_cli.cli_models import e2e_submit_results
-from electionguard_cli.cli_models import import_ballots
-from electionguard_cli.cli_models import shared
 
-from electionguard_cli.cli_models.e2e import (
-    E2eInputs,
-    e2e_inputs,
+from electionguard_cli.cli_models.cli_election_inputs_base import (
+    CliElectionInputsBase,
 )
 from electionguard_cli.cli_models.e2e_build_election_results import (
     BuildElectionResults,
@@ -18,29 +15,14 @@ from electionguard_cli.cli_models.e2e_decrypt_results import (
 from electionguard_cli.cli_models.e2e_submit_results import (
     E2eSubmitResults,
 )
-from electionguard_cli.cli_models.import_ballots import (
-    ImportBallotInputs,
-    import_ballot_inputs,
-)
-from electionguard_cli.cli_models.shared import (
-    CliElectionInputsBase,
-    cli_election_inputs_base,
-)
 
 __all__ = [
     "BuildElectionResults",
     "CliElectionInputsBase",
     "E2eDecryptResults",
-    "E2eInputs",
     "E2eSubmitResults",
-    "ImportBallotInputs",
     "cli_election_inputs_base",
-    "e2e",
     "e2e_build_election_results",
     "e2e_decrypt_results",
-    "e2e_inputs",
     "e2e_submit_results",
-    "import_ballot_inputs",
-    "import_ballots",
-    "shared",
 ]
