--- src/electionguard_cli/cli_models/e2e/__init__.py
@@ -1,7 +1,0 @@
-from electionguard_cli.cli_models.e2e import e2e_inputs
-
-from electionguard_cli.cli_models.e2e.e2e_inputs import (
-    E2eInputs,
-)
-
-__all__ = ["E2eInputs", "e2e_inputs"]
