--- src/electionguard_cli/cli_models/import_ballots/__init__.py
@@ -1,7 +1,0 @@
-from electionguard_cli.cli_models.import_ballots import import_ballot_inputs
-
-from electionguard_cli.cli_models.import_ballots.import_ballot_inputs import (
-    ImportBallotInputs,
-)
-
-__all__ = ["ImportBallotInputs", "import_ballot_inputs"]
