--- src/electionguard_cli/cli_models/shared/__init__.py
@@ -1,7 +1,0 @@
-from electionguard_cli.cli_models.shared import cli_election_inputs_base
-
-from electionguard_cli.cli_models.shared.cli_election_inputs_base import (
-    CliElectionInputsBase,
-)
-
-__all__ = ["CliElectionInputsBase", "cli_election_inputs_base"]
