+++ src/electionguard_cli/e2e/e2e_input_retrieval_step.py
@@ -7,10 +7,11 @@ from electionguard.manifest import Manifest
 from electionguard.serialize import (
     from_list_in_file_wrapper,
 )
-from electionguard_cli.cli_models.e2e.e2e_inputs import E2eInputs
-from electionguard_cli.steps.shared.input_retrieval_step_base import (
+
+from ..cli_steps import (
     InputRetrievalStepBase,
 )
+from .e2e_inputs import E2eInputs
 
 
 class E2eInputRetrievalStep(InputRetrievalStepBase):
