+++ src/electionguard_cli/e2e/e2e_inputs.py
@@ -2,7 +2,7 @@ from typing import List
 from electionguard.ballot import PlaintextBallot
 from electionguard.guardian import Guardian
 from electionguard.manifest import Manifest
-from electionguard_cli.cli_models.shared import (
+from ..cli_models import (
     CliElectionInputsBase,
 )
