--- src/electionguard_cli/steps/__init__.py
@@ -1,60 +1,0 @@
-from electionguard_cli.steps import e2e
-from electionguard_cli.steps import import_ballots
-from electionguard_cli.steps import shared
-
-from electionguard_cli.steps.e2e import (
-    E2eInputRetrievalStep,
-    KeyCeremonyStep,
-    SubmitVotesStep,
-    e2e_input_retrieval_step,
-    key_ceremony_step,
-    submit_votes_step,
-)
-from electionguard_cli.steps.import_ballots import (
-    ImportBallotsInputRetrievalStep,
-    import_ballots_input_retrieval_step,
-)
-from electionguard_cli.steps.shared import (
-    CliStepBase,
-    DecryptStep,
-    ElectionBuilderStep,
-    ElectionRecordStep,
-    InputRetrievalStepBase,
-    PrintResultsStep,
-    TallyStep,
-    cli_step_base,
-    decrypt_step,
-    election_builder_step,
-    election_record_step,
-    input_retrieval_step_base,
-    print_results_step,
-    tally_step,
-)
-
-__all__ = [
-    "CliStepBase",
-    "DecryptStep",
-    "E2eInputRetrievalStep",
-    "ElectionBuilderStep",
-    "ElectionRecordStep",
-    "ImportBallotsInputRetrievalStep",
-    "InputRetrievalStepBase",
-    "KeyCeremonyStep",
-    "PrintResultsStep",
-    "SubmitVotesStep",
-    "TallyStep",
-    "cli_step_base",
-    "decrypt_step",
-    "e2e",
-    "e2e_input_retrieval_step",
-    "election_builder_step",
-    "election_record_step",
-    "import_ballots",
-    "import_ballots_input_retrieval_step",
-    "input_retrieval_step_base",
-    "key_ceremony_step",
-    "print_results_step",
-    "shared",
-    "submit_votes_step",
-    "tally_step",
-]
