--- src/electionguard_cli/steps/e2e/__init__.py
@@ -1,22 +1,0 @@
-from electionguard_cli.steps.e2e import e2e_input_retrieval_step
-from electionguard_cli.steps.e2e import key_ceremony_step
-from electionguard_cli.steps.e2e import submit_votes_step
-
-from electionguard_cli.steps.e2e.e2e_input_retrieval_step import (
-    E2eInputRetrievalStep,
-)
-from electionguard_cli.steps.e2e.key_ceremony_step import (
-    KeyCeremonyStep,
-)
-from electionguard_cli.steps.e2e.submit_votes_step import (
-    SubmitVotesStep,
-)
-
-__all__ = [
-    "E2eInputRetrievalStep",
-    "KeyCeremonyStep",
-    "SubmitVotesStep",
-    "e2e_input_retrieval_step",
-    "key_ceremony_step",
-    "submit_votes_step",
-]
