--- src/electionguard_cli/steps/import_ballots/__init__.py
@@ -1,7 +1,0 @@
-from electionguard_cli.steps.import_ballots import import_ballots_input_retrieval_step
-
-from electionguard_cli.steps.import_ballots.import_ballots_input_retrieval_step import (
-    ImportBallotsInputRetrievalStep,
-)
-
-__all__ = ["ImportBallotsInputRetrievalStep", "import_ballots_input_retrieval_step"]
