--- src/electionguard_cli/steps/shared/__init__.py
@@ -1,46 +1,0 @@
-from electionguard_cli.steps.shared import cli_step_base
-from electionguard_cli.steps.shared import decrypt_step
-from electionguard_cli.steps.shared import election_builder_step
-from electionguard_cli.steps.shared import election_record_step
-from electionguard_cli.steps.shared import input_retrieval_step_base
-from electionguard_cli.steps.shared import print_results_step
-from electionguard_cli.steps.shared import tally_step
-
-from electionguard_cli.steps.shared.cli_step_base import (
-    CliStepBase,
-)
-from electionguard_cli.steps.shared.decrypt_step import (
-    DecryptStep,
-)
-from electionguard_cli.steps.shared.election_builder_step import (
-    ElectionBuilderStep,
-)
-from electionguard_cli.steps.shared.election_record_step import (
-    ElectionRecordStep,
-)
-from electionguard_cli.steps.shared.input_retrieval_step_base import (
-    InputRetrievalStepBase,
-)
-from electionguard_cli.steps.shared.print_results_step import (
-    PrintResultsStep,
-)
-from electionguard_cli.steps.shared.tally_step import (
-    TallyStep,
-)
-
-__all__ = [
-    "CliStepBase",
-    "DecryptStep",
-    "ElectionBuilderStep",
-    "ElectionRecordStep",
-    "InputRetrievalStepBase",
-    "PrintResultsStep",
-    "TallyStep",
-    "cli_step_base",
-    "decrypt_step",
-    "election_builder_step",
-    "election_record_step",
-    "input_retrieval_step_base",
-    "print_results_step",
-    "tally_step",
-]
